feat: cost per proof done, in wei

+++ explorer/priv/repo/migrations/20240711163410_add_cost_per_proof_column_in_batches.exs
@@ -3,7 +3,7 @@ defmodule Explorer.Repo.Migrations.AddCostPerProofColumnInBatches do
 
   def change do
     alter table("batches") do # add new columns
-      add :cost_per_proof, :integer
+      add :cost_per_proof, :bigint
     end
   end
 end
